changed errors to be thrown instead of console logged and moved database testing code into its own file
